third_party/blink/web_tests/media/media-fragments/TC0039-expected.txt


Title: TC0039
Fragment: 't=npt%3A3'
Comment: The media is requested from 3 seconds to the end. Note: %3A is equivalent to :.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
